cleaning (#247)
