+++ tfac/figures/figure3.py
@@ -1,29 +1,15 @@
-""" factorize the ECM data. """
-import pandas as pd
-import numpy as np
-import tensorly as tl
-from tensorly.cp_tensor import cp_flip_sign
-from tensorly.decomposition import parafac
-from tensorpack.cmtf import cp_normalize
-import seaborn as sns
-from ..dataHelpers import import_LINCS_MEMA, reorder_table
+""" factorize the MCF10A ECM data. """
+
+from ..dataHelpers import import_LINCS_MEMA
+from ..plotHelpers import plot_heatmaps
 from .common import getSetup
 
 
 def makeFigure():
     """ make heatmaps of factors when decomposed individually. """
+
     ax, f = getSetup((30, 15), (3, 1))
     tensor = import_LINCS_MEMA("mcf10a_ssc_Level4.tsv.xz")
-    fac = parafac(tensor.to_numpy(), 5, n_iter_max=2000, linesearch=True, tol=1e-9)
-    fac = cp_flip_sign(fac, 2)
-    fac = cp_normalize(fac)
-
-    labels = [f"Cmp. {i}" for i in np.arange(1, fac.factors[0].shape[1] + 1)]
-    print("R2X: ", 1.0 - np.linalg.norm(tl.cp_to_tensor(fac) - tensor)**2.0 / np.linalg.norm(tensor)**2.0)
-
-    for ii in range(tensor.ndim):
-        facZero = pd.DataFrame(fac.factors[ii], columns=labels, index=tensor.coords[tensor.dims[ii]])
-        facZero = reorder_table(facZero)
-        sns.heatmap(facZero.T, ax=ax[ii], cmap="PRGn", center=0)
+    plot_heatmaps(tensor, ax)
 
     return f
